- Proper unique outlines, still some debugging but functional currentlt

+++ shaders/phong.frag
@@ -4,11 +4,11 @@ in vec4 colourInterp;
 in vec3 viewNormal;
 in vec3 viewLightDirection;
 
-layout( location = 0 ) out vec4  colourOutput;         // GL_COLOR_ATTACHMENT0
-layout( location = 1 ) out vec4  dummy0;               // GL_COLOR_ATTACHMENT1 // outline
-layout( location = 2 ) out vec4  transparencyOutput;   // GL_COLOR_ATTACHMENT2
-layout( location = 3 ) out vec4  dummy2;      	       // GL_COLOR_ATTACHMENT3 // arthro
-layout( location = 4 ) out ivec3 pickingOutput;        // GL_COLOR_ATTACHMENT4
+layout( location = 0 ) out vec4   colourOutput;        // GL_COLOR_ATTACHMENT0
+layout( location = 1 ) out ivec4  dummy0;              // GL_COLOR_ATTACHMENT1 // outline
+layout( location = 2 ) out vec4   transparencyOutput;  // GL_COLOR_ATTACHMENT2
+layout( location = 3 ) out vec4   dummy2;      	       // GL_COLOR_ATTACHMENT3 // arthro
+layout( location = 4 ) out ivec3  pickingOutput;       // GL_COLOR_ATTACHMENT4
 
 uniform ivec3 uPickingColour;
